requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2==2.8.6
django-phonenumber-field[phonenumberslite]==7.1.0
drf-yasg==1.21.7